+++ src/components/Widgets/FeedbackWidget/index.js
@@ -1,7 +1,6 @@
 import { FeedbackProvider, useFeedbackContext } from './context';
 import useFeedbackData from './useFeedbackData';
 import FeedbackForm from './FeedbackForm';
-import FeedbackButton from './FeedbackButton';
 import FeedbackContainer from './FeedbackContainer';
 
-export { FeedbackProvider, useFeedbackContext, useFeedbackData, FeedbackForm, FeedbackButton, FeedbackContainer };
+export { FeedbackProvider, useFeedbackContext, useFeedbackData, FeedbackForm, FeedbackContainer };
